corrigindo um erro no delete

+++ src/services/CourseServices/DeleteCourseService.ts
@@ -7,6 +7,12 @@ class DeleteCourseService {
   async execute(id: string) {
     const courseRepository = getCustomRepository(CourseRepository)
 
+    const idExists = await courseRepository.findOne(id)
+
+    if (!idExists) {
+      throw new Error("This course doesn't exists!")
+    }
+
     const deleteCourse = await courseRepository.delete(id)
 
     return deleteCourse
