+++ src/services/UserServices/DeleteUserService.ts
@@ -7,6 +7,12 @@ class DeleteUserService {
   async execute(id: string) {
     const userRepository = getCustomRepository(UserRepository)
 
+    const idExists = await userRepository.findOne(id)
+
+    if (!idExists) {
+      throw new Error("This user doesn't exists!")
+    }
+
     const deleteUser = await userRepository.delete(id)
 
     return deleteUser
